feat: multi aim, two bone IK constraint demo
This commit adds two bone IK and multi-aim constraints to the male
character rig. This allows for more realistic and dynamic movement
of the character's limbs and head.

The two bone IK constraint is used to control the position of the
character's arms and legs, while the multi-aim constraint is used
to control the direction of the character's head.

reference tutorial video: https://youtu.be/Yf0th6ALiFA?si=o6CQxqwno_JPihzf